+++ src/Files.App/Dialogs/CreateShortcutDialog.xaml.cs
@@ -8,6 +8,9 @@ using Microsoft.UI.Xaml.Data;
 
 namespace Files.App.Dialogs
 {
+	/// <summary>
+	/// Represents an <see cref="ContentDialog"/> UI for shortcut creation.
+	/// </summary>
 	public sealed partial class CreateShortcutDialog : ContentDialog, IDialog<CreateShortcutDialogViewModel>
 	{
 		public CreateShortcutDialogViewModel ViewModel { get;  set; }
