+++ src/Files.App/Dialogs/DecompressArchiveDialog.xaml.cs
@@ -8,6 +8,9 @@ using System.Text;
 
 namespace Files.App.Dialogs
 {
+	/// <summary>
+	/// Represents an <see cref="ContentDialog"/> UI for archive decompression.
+	/// </summary>
 	public sealed partial class DecompressArchiveDialog : ContentDialog
 	{
 		public DecompressArchiveDialogViewModel ViewModel { get; set; }
@@ -19,6 +22,7 @@ namespace Files.App.Dialogs
 
 		private void ContentDialog_PrimaryButtonClick(ContentDialog sender, ContentDialogButtonClickEventArgs args)
 		{
+			// TODO: Do not manually invoke a command
 			if (ViewModel.IsArchiveEncrypted)
 				ViewModel.PrimaryButtonClickCommand.Execute(new DisposableArray(Encoding.UTF8.GetBytes(Password.Password)));
 		}
