+++ src/Files.App/Dialogs/ElevateConfirmDialog.xaml.cs
@@ -6,6 +6,9 @@ using Microsoft.UI.Xaml.Controls;
 
 namespace Files.App.Dialogs
 {
+	/// <summary>
+	/// Represents an <see cref="ContentDialog"/> UI for elevating confirmation.
+	/// </summary>
 	public sealed partial class ElevateConfirmDialog : ContentDialog, IDialog<ElevateConfirmDialogViewModel>
 	{
 		public ElevateConfirmDialogViewModel ViewModel { get; set; }
@@ -17,7 +20,7 @@ namespace Files.App.Dialogs
 
 		public new async Task<DialogResult> ShowAsync()
 		{
-			return await base.ShowAsync();
+			return (DialogResult)await base.ShowAsync();
 		}
 	}
 }
